+++ src/Illuminate/Http/Controller/Data/GetDataEntitiesController.php
@@ -16,8 +16,6 @@ use Illuminate\Http\Request;
  * to fetch data based on user authentication and query parameters.
  *
  * @package AppTank\Horus\Illuminate\Http\Controller
- *
- * @deprecated This class is deprecated and will be removed in future versions. Because its uses is unperformant.
  */
 class GetDataEntitiesController extends Controller
 {